+++ services/centrum/housekeeper/dispatches.go
@@ -157,10 +157,10 @@ func (s *Housekeeper) cancelOldDispatches(ctx context.Context) error {
 				),
 			),
 			tDispatchStatus.Status.NOT_IN(
-				jet.Int16(int16(centrum.StatusDispatch_STATUS_DISPATCH_COMPLETED)),
-				jet.Int16(int16(centrum.StatusDispatch_STATUS_DISPATCH_CANCELLED)),
-				jet.Int16(int16(centrum.StatusDispatch_STATUS_DISPATCH_ARCHIVED)),
-				jet.Int16(int16(centrum.StatusDispatch_STATUS_DISPATCH_COMPLETED)),
+				jet.Int32(int32(centrum.StatusDispatch_STATUS_DISPATCH_COMPLETED)),
+				jet.Int32(int32(centrum.StatusDispatch_STATUS_DISPATCH_CANCELLED)),
+				jet.Int32(int32(centrum.StatusDispatch_STATUS_DISPATCH_ARCHIVED)),
+				jet.Int32(int32(centrum.StatusDispatch_STATUS_DISPATCH_COMPLETED)),
 			),
 			tDispatch.CreatedAt.LT_EQ(
 				jet.CURRENT_TIMESTAMP().SUB(jet.INTERVAL(60, jet.MINUTE)),
